Move reconnection test to scout tests
